Migrate invite_2a_claimer tests
