+++ libparsec/crates/protocol/tests/invite.rs
@@ -136,98 +136,6 @@ fn serde_invite_info_rep(#[case] raw: &[u8], #[case] expected: invited_cmds::inv
 }
 
 #[parsec_test]
-fn serde_invite_2a_claimer_send_hashed_nonce_req() {
-    // Generated from Python implementation (Parsec v2.10.0+dev)
-    // Content:
-    //   claimer_hashed_nonce: hex!("e37ce3b00a1f15b3de62029972345420b76313a885c6ccc6e3b5547857b3ecc6")
-    //   cmd: "invite_2a_claimer_send_hashed_nonce"
-    let raw = hex!(
-        "82b4636c61696d65725f6861736865645f6e6f6e6365c420e37ce3b00a1f15b3de62029972"
-        "345420b76313a885c6ccc6e3b5547857b3ecc6a3636d64d923696e766974655f32615f636c"
-        "61696d65725f73656e645f6861736865645f6e6f6e6365"
-    );
-
-    let req = invited_cmds::invite_2a_claimer_send_hashed_nonce::Req {
-        claimer_hashed_nonce: HashDigest::from(hex!(
-            "e37ce3b00a1f15b3de62029972345420b76313a885c6ccc6e3b5547857b3ecc6"
-        )),
-    };
-
-    let expected = invited_cmds::AnyCmdReq::Invite2aClaimerSendHashedNonce(req);
-
-    let data = invited_cmds::AnyCmdReq::load(&raw).unwrap();
-
-    assert_eq!(data, expected);
-
-    // Also test serialization round trip
-    let raw2 = if let invited_cmds::AnyCmdReq::Invite2aClaimerSendHashedNonce(data) = data {
-        data.dump().unwrap()
-    } else {
-        unreachable!()
-    };
-
-    let data2 = invited_cmds::AnyCmdReq::load(&raw2).unwrap();
-
-    assert_eq!(data2, expected);
-}
-
-#[parsec_test]
-#[case::ok(
-    // Generated from Python implementation (Parsec v2.6.0+dev)
-    // Content:
-    //   greeter_nonce: hex!("666f6f626172")
-    //   status: "ok"
-    &hex!(
-        "82ad677265657465725f6e6f6e6365c406666f6f626172a6737461747573a26f6b"
-    )[..],
-    invited_cmds::invite_2a_claimer_send_hashed_nonce::Rep::Ok {
-        greeter_nonce: b"foobar".as_ref().into(),
-    }
-)]
-#[case::not_found(
-    // Generated from Python implementation (Parsec v2.6.0+dev)
-    // Content:
-    //   status: "not_found"
-    &hex!(
-        "81a6737461747573a96e6f745f666f756e64"
-    )[..],
-    invited_cmds::invite_2a_claimer_send_hashed_nonce::Rep::NotFound
-)]
-#[case::already_deleted(
-    // Generated from Python implementation (Parsec v2.6.0+dev)
-    // Content:
-    //   status: "already_deleted"
-    &hex!(
-        "81a6737461747573af616c72656164795f64656c65746564"
-    )[..],
-    invited_cmds::invite_2a_claimer_send_hashed_nonce::Rep::AlreadyDeleted
-)]
-#[case::invalid_state(
-    // Generated from Python implementation (Parsec v2.6.0+dev)
-    // Content:
-    //   status: "invalid_state"
-    &hex!(
-        "81a6737461747573ad696e76616c69645f7374617465"
-    )[..],
-    invited_cmds::invite_2a_claimer_send_hashed_nonce::Rep::InvalidState
-)]
-fn serde_invite_2a_claimer_send_hashed_nonce_rep(
-    #[case] raw: &[u8],
-    #[case] expected: invited_cmds::invite_2a_claimer_send_hashed_nonce::Rep,
-) {
-    let data = invited_cmds::invite_2a_claimer_send_hashed_nonce::Rep::load(raw).unwrap();
-
-    assert_eq!(data, expected);
-
-    // Also test serialization round trip
-    let raw2 = data.dump().unwrap();
-
-    let data2 = invited_cmds::invite_2a_claimer_send_hashed_nonce::Rep::load(&raw2).unwrap();
-
-    assert_eq!(data2, expected);
-}
-
-#[parsec_test]
 fn serde_invite_2b_greeter_send_nonce_req() {
     // Generated from Python implementation (Parsec v2.6.0+dev)
     // Content:
